Added ability to stop task, and delete files
Added the ability for the end user to stop the write/delete of files
with a button.

Added the function allowing the application to delete files from the
directory is just wrote to.

Set the window size.

Updated the Progress Bar to better reflect the actual progress that has
been made.